Add tests in the repository

+++ GraphRanker.c
@@ -7,9 +7,11 @@
 #define RED 'r'
 #define BLACK 'b'
 
-int graphReadBefore = 0;
-int insertions = 0;
-long int numberOfNodes, topGraph;
+int graphReadBefore = 0;    //name of graphs
+int insertions = 0;         //determine if there will be a check in the tree
+int firstPrint = 0;
+int changed = 1;
+long int numberOfNodes, topGraph;   //input, numberOfNodes are the nodes in the graph, topGraph the number of graph saved
 char row[MAX];
 
 typedef struct listNode{
@@ -44,6 +46,7 @@ int main(){
     char *tmp;
     int i, k = 0 ;
     node_t *root = NIL;
+    node_t *nodeToDelete = NIL;
 
     tmp = fgets(input, MAX, stdin);
     for(i = 0; i < MAX; i++){
@@ -63,33 +66,57 @@ int main(){
     topGraph = strtol(converter, &tmp, 0);
     while(NULL != fgets(input, MAX, stdin)){
         if(0 == strncmp("A", input, 1)){
-            //printf("\n\nCHECKING THE GRAPH: %d\n", graphReadBefore);
             long long int sum = takeGraph(numberOfNodes);
-            addGraph(&root, sum);
             if(insertions >= topGraph){
-                node_t *nodeToDelete = root;
-                while(NIL != nodeToDelete->right || NIL != nodeToDelete->left){
-                    if(nodeToDelete->right->valueOfThePath > nodeToDelete->valueOfThePath){
-                        nodeToDelete = nodeToDelete->right;
-                    }
-                    else if(nodeToDelete->left->valueOfThePath > nodeToDelete->valueOfThePath){
-                        nodeToDelete = nodeToDelete->left;
-                    }
-                    else{
-                        break;
+                //find the node with the bigger graph
+                if(1 == changed){
+                    nodeToDelete = root;
+                    while(NIL != nodeToDelete->right || NIL != nodeToDelete->left){
+                        if(nodeToDelete->right->valueOfThePath > nodeToDelete->valueOfThePath){
+                            nodeToDelete = nodeToDelete->right;
+                        }
+                        else if(nodeToDelete->left->valueOfThePath > nodeToDelete->valueOfThePath){
+                            nodeToDelete = nodeToDelete->left;
+                        }
+                        else if(nodeToDelete->valueOfThePath == nodeToDelete->right->valueOfThePath &&
+                                nodeToDelete->valueOfThePath == nodeToDelete->left->valueOfThePath){
+                            //the graphs have the same value of the path
+                            if(nodeToDelete->name < nodeToDelete->right->name){
+                                nodeToDelete = nodeToDelete->right;
+                            }
+                            else{
+                                nodeToDelete = nodeToDelete->left;
+                            }
+                        }
+                        else{
+                            break;
+                        }
                     }
                 }
-                deleteGraph(&root, nodeToDelete);
+                //if the new graph is smaller it will add to the tree, than eliminate the biggest
+                if(nodeToDelete->valueOfThePath > sum){
+                    addGraph(&root, sum);
+                    deleteGraph(&root, nodeToDelete);
+                    changed = 1;
+                }
+                //if the new graph is not smaller than do nothing and the next iteration don't nedd to find the biggest graph
+                else{
+                    changed = 0;
+                }
+            }
+            //add graph till k insertions
+            else{
+                addGraph(&root, sum);
             }
             graphReadBefore++;
             insertions++;
         }
         if(0 == strncmp("T", input, 1)){
             TopK(&root, topGraph);
+            firstPrint = 1;
             printf("\n");
         }
     }
-    //printf("\n");
     return 0;
 }
 
@@ -100,22 +127,25 @@ int main(){
 long long int takeGraph(long int nodes){
     long int graphMatrix[nodes][nodes];
     char *inputDist;
-    int i, j, best = 1;
+    int i, j, best = 1, firstLine = 1;
     for(i = 0; i < nodes; i++){
         inputDist = fgets(row, MAX, stdin);
         for(j = 0; j < nodes; j++){
             graphMatrix[i][j] = strtol(inputDist, &inputDist, 10);
             inputDist++;
         }
-    }
-    for(int k = 1; k < nodes; k++){
-        if(0 != graphMatrix[0][k]){
-            best = 0;
+        //check if the first line has only zeroes (except for the [0][0]), if it's than the other lines are useless
+        if(1 == firstLine){
+            for(int k = 1; k < nodes; k++){
+                if(0 != graphMatrix[0][k]){
+                    best = 0;
+                }
+            }
+            if(1 == best) {
+                return 0;
+            }
+            firstLine = 0;
         }
-    }
-    if(1 == best) {
-        //printf("The minimum path of this graph is:0\n");
-        return 0;
     }
     return dijkstra(0, graphMatrix);
 }
@@ -187,7 +217,6 @@ long long int sumPaths(const long int distancesFrom0[]){
     for(int i = 1; i < numberOfNodes; i++){
         total = total + distancesFrom0[i];
     }
-    //printf("The minimum path of this graph is:%lld\n", total);
     return total;
 }
 
@@ -195,18 +224,15 @@ long long int sumPaths(const long int distancesFrom0[]){
  * add the graph given by the input
 */
 void addGraph(node_t **root, long long int pathDistance){
-    //printf("addGraph started\n");
     node_t *y = NIL;
     node_t *x = *root;
-    //fin the last node useful
+    //find the last node useful
     while(x != NIL){
         y = x;
         if(pathDistance < x->valueOfThePath){
-            //printf("Go left\n");
             x = x->left;
         }
         else{
-            //printf("Go right\n");
             x = x->right;
         }
     }
@@ -218,75 +244,70 @@ void addGraph(node_t **root, long long int pathDistance){
     newNode->right = NIL;
     newNode->color = RED;
     if(NIL == y){
-        //printf("Y is NIL so it's the first graph to be added\n");
         *root = newNode;
     }
     else{
         if(newNode->valueOfThePath < y->valueOfThePath){
-            //printf("The new graph is smaller, go left\n");
             y->left = newNode;
         }
         else{
-            //printf("The new graph is bigger, go right\n");
             y->right = newNode;
         }
     }
     insertFixup(root, newNode);
-    //printf("The root is the graph %d\n", (*root)->name);
 }
 
 /**
  * function to balance the tree after an insertion of a node
 */
 void insertFixup(node_t **root, node_t *z){
-    //printf("insterFixup started\n");
     while (z != (*root) && z->parent->color == RED){
         node_t *x = z->parent;
         if (x->color == RED){
-            //printf("The parent is red\n");
+            //the parent is red
             if (x == x->parent->left){
-                //printf("The parent is a left child\n");
+                //the parent is a left child
                 node_t *y = x->parent->right;
                 if (y->color == RED){
-                    //printf("The uncle is red, switch colors and the hew node became it's grand parent\n");
+                    //the uncle is red, switch colors and the hew node became it's grand parent
                     x->color = BLACK;
                     y->color = BLACK;
                     x->parent->color = RED;
                     z = x->parent;
                 }
                 else{
-                    //printf("The uncle is black\n");
+                    //the uncle is black
                     if (z == x->right){
-                        //printf("The new node is a right child, need a left rotation\n");
+                        //the new node is a right child, need a left rotation
                         z = x;
                         leftRotate(root, z);
                         x = z->parent;
                     }
-                    //printf("switch colors and right rotation\n");
+                    //switch colors and right rotation
                     x->color = BLACK;
                     x->parent->color = RED;
                     rightRotate(root, x->parent);
                 }
             }
             else{
-                //printf("The parent is a right child\n");
+                //the parent is a right child
                 node_t *y = x->parent->left;
                 if (y->color == RED){
-                    //printf("The uncle is red, switch colors and the hew node became it's grand parent\n");
+                    //the uncle is red, switch colors and the hew node became it's grand parent
                     x->color = BLACK;
                     y->color = BLACK;
                     x->parent->color = RED;
                     z = x->parent;
                 }
                 else{
-                    //printf("The uncle is black\n");
+                    //the uncle is black
                     if (z == x->left){
-                        //printf("The new node is a left child, need a right rotation\n");
+                        //the new node is a left child, need a right rotation
                         z = x;
                         rightRotate(root, z);
                         x = z->parent;
                     }
-                    //printf("switch colors and left rotation\n");
+                    //switch colors and left rotation
                     x->color = BLACK;
                     x->parent->color = RED;
                     leftRotate(root, x->parent);
@@ -299,7 +320,6 @@ void insertFixup(node_t **root, node_t *z){
 
 /**
  * delete the node (graph) with the maximum path distance
- * (the node at the most right in the tree)
 */
 void deleteGraph(node_t **root, node_t *z){
     node_t *y = NIL;
@@ -335,7 +355,6 @@ void deleteGraph(node_t **root, node_t *z){
     if(BLACK == y->color){
         deleteFixup(root, x);
     }
-    //printf("Eliminated graph %d\n", y->name);
     free(y);
 }
 
@@ -359,16 +378,14 @@ node_t *treeMinimum(node_t *x){
 }
 
 /**
- * function to balance the tree after an erasure of a node
+ * function to balance the tree after an elimination of a node
 */
 void deleteFixup(node_t **root, node_t *x){
-    //printf("Delete fixup started\n");
     while (x != (*root) && x->color == BLACK){
         if (x == x->parent->left){
             //if the node is a left child
             node_t *w = x->parent->right;
             if (w->color == RED){
-                //printf("The parent is red\n");
                 //if w is red, need recolor and a left rotation
                 w->color = BLACK;
                 x->parent->color = RED;
@@ -376,22 +393,20 @@ void deleteFixup(node_t **root, node_t *x){
                 w = x->parent->right;
             }
             if (w->left->color == BLACK && w->right->color == BLACK){
-                //printf("The uncle is red, switch colors and the hew node became it's grand parent\n");
                 //if w has two black children it has to be red
                 w->color = RED;
                 x = x->parent;
             }
             else{
-                //printf("The uncle is black\n");
+                //the uncle is black
                 if (w->right->color == BLACK){
-                    //printf("The new node is a left child, need a right rotation\n");
                     //w has the right black child, recolor and right rotation
                     w->left->color = BLACK;
                     w->color = RED;
                     rightRotate(root, w);
                     w = x->parent->right;
                 }
-                //printf("switch colors and left rotation\n");
+                //switch colors and left rotation
                 w->color = x->parent->color;
                 x->parent->color = BLACK;
                 w->right->color = BLACK;
@@ -400,11 +415,10 @@ void deleteFixup(node_t **root, node_t *x){
             }
         }
         else{
-            //printf("The parent is black\n");
             //same but with right and left changed
             node_t *w = x->parent->left;
             if (w->color == RED){
-                //printf("The uncle is red, switch colors and the hew node became it's grand parent\n");
+                //the uncle is red, switch colors and the hew node became it's grand parent
                 w->color = BLACK;
                 x->parent->color = RED;
                 rightRotate(root, x->parent);
@@ -433,7 +447,7 @@ void deleteFixup(node_t **root, node_t *x){
 }
 
 /**
- * Right and Left rotation of the nodes in the tree to balance it after an insertion or an erasure
+ * Right and Left rotation of the nodes in the tree to balance it after an insertion or an elimination
 */
 void leftRotate(node_t **root, node_t *x){
     node_t *y = x->right;
@@ -472,13 +486,20 @@ void rightRotate(node_t **root, node_t *x){
 }
 
 /**
- * inorder print names of the graphs
+ * inorder print by value of the path
+ * print the names of the graphs
 */
 void TopK(node_t **root, int k){
     if(NIL == *root){
         return;
     }
     TopK(&(*root)->left, k-1);
-    printf("%d " , (*root)->name);
+    if(0 == firstPrint){
+        printf("%d", (*root)->name);
+        firstPrint = 1;
+    }
+    else{
+        printf(" %d", (*root)->name);
+    }
     TopK(&(*root)->right, k-1);
 }
